feat: simulate a ⏰ counter mechanism
Changes:
- Tried to simulate a counter mechanism using combination of state and effect hooks.
- Starts and stops using buttons that change "canStart" state variable.

Implementation:
created two state hooks, which holds time:Number and canStart:Booleam, and a
useEffect hook. canStart is passed into the dependency array. The clean up function
resets the timer to 0 and clears out the interval created. When mounted again
with canStart as false, the state of time won't change due to condition. and the
interval is cleared out hence stoping re-renders.